Lay out c-project-tile__header
I want the subtitle (which can be the document handle, or a generic
label like "Software") to appear above the product title. But in the
source order, it semantically makes more sense for the <h2> to appear
first. Here I'm using flexbox to swap the order for the __title and
__subtitle elements in the c-project-tile__header component.