Populate homepage tiles with MongoDB query
Use a MongoDB query to populate the MVP homepage instead of using
document-demo.yaml.

The product-tile Jinja template was adjusted to natively use the schema
terms from MongoDB rather than having a transformation. We'll have to
see if this is sustainable, or if we want to do a pre-template
document transformation.