Convert description to HTML <p> with Markdown
The technote description is typically plain text, but I do want it to be
marked up as a paragraph element in HTML. It could also be multiple
paragraphs. To avoid having to parse these description paragraphs myself
and render HTML, I'm opting to use a markdown processor. Note that
markdown passes through HTML, which is ideal. It also marks up
paragraphs.

I'm using misaka==2.1.0 as the markdown processor and wrapped it as a
'markdown' Jinja filter.