Create GitHub Actions CI workflow
This workflow has two jobs: lint the code and ensure the Gatsby site
builds.

We're pinning to use node v13.3.0 for build. This matches my local node
version and is known to be compatible with the package-lock.json. I find
that if we float the version of node then sometimes the build will break
on us.