Use flexbox layout for searchbox+poweredby
Under the community license we need to use the "PoweredBy" component to
give credit to Algolia. A nice way to integrate that into the layout is
to have the PoweredBy component side next to the searchbox. Flexbox lets
the searchbox take up as much space as possible, while ensuring that the
Algolia logo stays on that line.