Experiment using details for card metadata
Ultimately we want to put details about a record, like abstract,
authors, date updated, etc. in a collapsed section. This keeps the
initial search compact, but still lets people get the metadata from the
search page without clicking through to the record.

This implementation using HTML5 details and summary. The drawbacks here
are:

1. Doesn't yet update the "details" text whether its expanded or not.
   Ideally I want to do a "Show more" / "Show less" button and have the
   "Show less" button float below the content.

2. It doesn't animate on open/close.

I think that ultimately we'll have to build a custom component for this.
For now, <details> works for prototyping.