Add DetailsToggleButton
In the AdvancedSearchPage component we create state using the useState
React hook that controls whether all the cards are expanded or not.
This state is toggled by the DetailsToggleButton. That state is also
drilled into the custom Hits component and the DocumentHit itself to set
the open attribute on the <details> toggle.