Use HierarchicalMenu for content type refinement
On the Advanced search page we can have all types of content. It
therefore makes sense to use a hierarchical refinement to let a use
provide specificity on what type of content they'd like to see.

This works with the "contentCategories.lxlX" record field added in
DM-24241.

Note that these record attributes need to be maked as attributes for
faceting in the Algolia config UI.