Add AppThemeProvider for theming
We're using the React Context API to create a context for UI theming
with styled-components. (e.g. dark or light color schemes). This context
holds state and functions for manipulating state.

The AppThemeProvider is wraps the root element via the wrapRootElement
hook in gatsby-browser.js. Thus all pages/components are descendents on
the theme provider and its context.

The purpose of the AppThemeProvider is to hold functions for
manipulating state. Other components can access those function via
useTheme. Additionally, the AppThemeProvider contains the
styled-components ThemeProvider. The ThemeProvider specifically contains
the UI theme state. The idea for the AppThemeProvider comes from the
rossbulat dark mode Medium article (see below).

Reference material that helped with this commit:

- https://medium.com/@rossbulat/react-dark-mode-with-styled-theming-and-context-57557de6400
- https://github.com/rossbulat/react-theming-dark-mode
- https://medium.com/@rossbulat/creating-themes-in-react-with-styled-components-6fce744b4e54
- https://www.gatsbyjs.org/blog/2019-01-31-using-react-context-api-with-gatsby/
- https://reactjs.org/docs/context.html