Add AutoSortBy index select component
The purpose of this component is to automatically toggle from a sorted
search refined (e.g. by ID or date) to a relevance sorting when a person
starts typing a search query.

The control is entirely over the default selection. That way, if the
user decides to override the sorting, they'll keep that selection.