Demonstrate sourcing Guides data from YAML
Information about guides is stored in a single YAML file
(data/guides.yaml). With gatsby-transformer-yaml we're able to get that
YAML data into the Gatsby data layer (GraphQl). In the FeaturedGuides
component we perform a GraphQl of the guides with the
"editorial/featured" tag and then list out the titles of those guides.

Next steps: Create a grid component and put individual results in cards