Create pages with collections of user guides
The templates/guideCollection.js template for pages that list user
guides as defined in src/data/guides.js

Actual pages are instantiated in gatsby-node.js using a GraphQL query of
the guideCollections.yaml file. That file defines collections of user
guides that share the same tag.

The rendering of guide cards in the guideCollection template pages is
handled by the GuideGrid component. GuideGrid takes a prop with the
guides that are part of the grid. We've refactored the FeaturedGrid to
use the GuideGrid component as the unique aspect of FeaturedGuides is
its static query for featured user guides.