Lay out page link with a NavGrid component
Now the links to document category pages, as well as new links to guide
collection pages, are laid out using a NavGrid component. The data for
what links are presented comes from GraphQL queries.

Because we can't have both a page-level query and a StaticQuery in the
same file, I moved the BackgroundSection component to its own file.