Update how eslint and prettier are managed
Eslint and prettier now run separately, which is the recommended method.
As part of this, we're dropping the usage of the wesbos eslint config
because that was stopping us from using eslint 7. I've ported in the
basic eslint config that we got from the wesbos config.

As well, instead of using pre-commit and Python to run git hooks, we're
using husky.