Switch to using internal component state
Rather than using algoliasearch as the state for the component, we're
now maintaining the component's state internally. This solves the
problem of the UI results bouncing rapidly when someone tries to
manipulate the text fields --- often resulting in it becoming impossible
to use the component.

Now we have an explicit submission button to send the component's new
value to algolia. Perhaps in the future we can change this to submit
after a delay of inactivity, or when the component loses focus? For now,
this is the most reliable interface.

Still needs work to style the button and form.