Add RefinementOptIn component
This component solves a problem where a refinement always applies if its
mounted. This is tricky when having a date-based filter, and certain
records are missing that filter's attribute --- they could never appear
in the search results. Thus we solve that by allowing a user to
explicitly opt-into using a filter that can exclude results if records
don't have the appropriate metadata.